Rename file and move send_outputs()
